pyproject.toml
[project]
name = "channel-point-discord"
version = "0.1.0"
description = "like a twitch channel point system in discord"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
]
